Import Polytope Studio lowpoly assets
Add Polytope Studio lowpoly asset pack and demos: imports Lowpoly_Characters, Modular_Armors (including PBR and Toon shaders), URP settings, and a comprehensive Weapons demo (prefabs, materials, textures, terrain, lighting and associated .meta files). Also modifies an existing Blink human prefab (HumanMale_Character_FREE.prefab). This brings art/shader/content needed for low-poly character/weapon visuals and demo scenes.

+++ Assets/_Scripts/MainGame/UI/GameOverUIController.cs
@@ -23,13 +23,13 @@ namespace _Scripts.MainGame.UI
                 restartButton.onClick.AddListener(RestartGame);
         }
 
-        public void Show(int score, string time, string difficulty)
+        public void Show(double score, string time, string difficulty)
         {
             if (contentPanel != null)
                 contentPanel.SetActive(true);
 
             if (scoreValueText != null)
-                scoreValueText.text = score.ToString();
+                scoreValueText.text = ((int)score).ToString(); // Still display as int
 
             if (timeValueText != null)
                 timeValueText.text = time;
